refactor(mentor): css추가 및 오류 수정
