+++ src/App.js
@@ -1,5 +1,9 @@
-import logo from "./logo.svg";
-import "./App.css";
+import "./styles/reset.css";
+//import logo from "./logo.svg";
+import "./styles/App.css";
+import Header from "./components/Header.js";
+import Main from "./components/Main.js";
+import Footer from "./components/Footer.js";
 
 function App() {
   return (
