+++ src/components/Main.js
@@ -1,0 +1,5 @@
+import "../styles/main.css";
+
+function Main() {}
+
+export default Main;
