+++ src/styles/header.css
@@ -1,21 +1,23 @@
 header {
+  height: 94px;
+
   display: flex;
   flex-direction: column;
   align-items: center;
-  width: 100%;
   gap: 8px;
-  background: var(--linkbrary-bg, #f0f6ff);
+
+  background: #f0f6ff;
 }
 
 .header__container {
   position: fixed;
+  width: 100%;
   z-index: 1;
+  height: 94px;
+
   display: flex;
   justify-content: space-between;
   align-items: center;
-  width: 100%;
-  max-width: 1920px;
-  height: 94px;
   padding: 0 200px;
 }
 
@@ -31,17 +33,34 @@ header {
 }
 
 button {
+  width: 128px;
+
+  display: flex;
+  padding: 16px 20px;
+  justify-content: center;
+  align-items: center;
+  gap: 10px;
+
+  border-radius: 8px;
+  background: var(
+    --gra-purpleblue-to-skyblue,
+    linear-gradient(91deg, #6d6afe 0.12%, #6ae3fe 101.84%)
+  );
+  text-decoration: none;
+  color: #ffffff;
+  font-size: 18px;
+  font-weight: 600;
 }
 
 .header__profile--logo {
   width: 10px;
   height: 10px;
+
   flex-shrink: 0;
 }
 
 .header__profile--email {
-  color: var(--linkbrary-gray-100, #373740);
-
+  color: #373740;
   /* Linkbrary/body2-regular */
   font-family: Pretendard;
   font-size: 14px;
